Input: Slight rewrite of the input system, moved to static accessing
In the future, we might want to switch to something like
```cpp
struct InputState {
    std::array<bool, OB_KEY_LAST + 1> current{};
    std::array<bool, OB_KEY_LAST + 1> previous{};
};
```
to allow for more complex commands at the same time, e.g. SUPER + ALT + G. This system works for now, but using a vector means we're not using O(1) whilst I do think that it would be possible to achieve O(1).